Use uint64_t for PTS and DTS fields
Timestamps in Matroska and other containers, as well as in GStreamer,
are defined as unsigned int 64 bit. Values larger than 2^63 can appear
in stress-tests.

This patch switches to uint64_t representation of timestamps for ffmpeg
parsers.

Also the list of constructors for TreeItem needs to be extended to
properly support various integer formats.